[~] Fixed any player can join to the server when all punishments are loaded.

+++ src/main/java/me/leoko/advancedban/manager/LogManager.java
@@ -32,10 +32,8 @@ public class LogManager {
         File[] fList = logsFolder.listFiles();
         // Auto purge for old logs
         for (File file : fList) {
-            if (file.isFile() && file.getName().contains(".gz")) {
-                if ((System.currentTimeMillis() - file.lastModified()) >= universal.getMethods().getInteger(universal.getMethods().getConfig(), "Log Purge Days") * 86400000L) {
-                    file.delete();
-                }
+            if (file.isFile() && file.getName().contains(".gz") && (System.currentTimeMillis() - file.lastModified()) >= universal.getMethods().getInteger(universal.getMethods().getConfig(), "Log Purge Days") * 86400000L) {
+                file.delete();
             }
         }
     }
